Removes initext_functions.cpp and merges its contents into initext_hooks.cpp.
